+++ crates/teloxide-core/src/payloads/send_game.rs
@@ -2,7 +2,7 @@
 
 use serde::Serialize;
 
-use crate::types::{ChatId, Message, ReplyMarkup, ReplyParameters, ThreadId};
+use crate::types::{BusinessConnectionId, ChatId, Message, ReplyMarkup, ReplyParameters, ThreadId};
 
 impl_payload! {
     /// Use this method to send a game. On success, the sent [`Message`] is returned.
@@ -18,7 +18,7 @@ impl_payload! {
         }
         optional {
             /// Unique identifier of the business connection on behalf of which the message will be sent
-            pub business_connection_id: String [into],
+            pub business_connection_id: BusinessConnectionId,
             /// Unique identifier for the target message thread (topic) of the forum; for forum supergroups only
             pub message_thread_id: ThreadId,
             /// Sends the message [silently]. Users will receive a notification with no sound.