+++ crates/teloxide-core/src/payloads/send_message.rs
@@ -3,8 +3,8 @@
 use serde::Serialize;
 
 use crate::types::{
-    LinkPreviewOptions, Message, MessageEntity, ParseMode, Recipient, ReplyMarkup, ReplyParameters,
-    ThreadId,
+    BusinessConnectionId, LinkPreviewOptions, Message, MessageEntity, ParseMode, Recipient,
+    ReplyMarkup, ReplyParameters, ThreadId,
 };
 
 impl_payload! {
@@ -21,7 +21,7 @@ impl_payload! {
         }
         optional {
             /// Unique identifier of the business connection on behalf of which the message will be sent
-            pub business_connection_id: String [into],
+            pub business_connection_id: BusinessConnectionId,
             /// Unique identifier for the target message thread (topic) of the forum; for forum supergroups only
             pub message_thread_id: ThreadId,
             /// Mode for parsing entities in the message text. See [formatting options] for more details.