+++ crates/teloxide-core/src/types/business_messages_deleted.rs
@@ -1,13 +1,13 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::{Chat, MessageId};
+use crate::types::{BusinessConnectionId, Chat, MessageId};
 
 /// This object is received when messages are deleted from a connected business
 /// account.
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct BusinessMessagesDeleted {
     /// Unique identifier of the business connection.
-    pub business_connection_id: String,
+    pub business_connection_id: BusinessConnectionId,
 
     /// Information about a chat in the business account. The bot may not have
     /// access to the chat or the corresponding user.